Get Gradient from CSS
