+++ MagicGradients/CssFormsGradientSource.cs
@@ -7,36 +7,14 @@ using MagicGradients;
 using MagicGradients.Parser;
 using Xamarin.Forms;
 using Xamarin.Forms.StyleSheets;
+using Xamarin.Forms.Xaml;
 
 namespace MagicGradients
 {
     [ContentProperty(nameof(Stylesheet))]
+    [TypeConverter(typeof(CssFormsGradientTypeConverter))]
     public class CssFormsGradientSource : BindableObject, ILinearGradientSource
     {
-        static CssFormsGradientSource()
-        {
-            var stylePropertyInfo = typeof(Xamarin.Forms.Internals.Registrar).GetProperty("StyleProperties",
-                BindingFlags.Static | BindingFlags.NonPublic);
-            if (stylePropertyInfo == null)
-                return;
-
-            var styleProperties = stylePropertyInfo.GetValue(null);
-
-            var styleAttributeType = typeof(StyleSheet).Assembly.GetType("Xamarin.Forms.StyleSheets.StylePropertyAttribute");
-            var styleAttributeInstance = Activator.CreateInstance(styleAttributeType, "gradient",
-                typeof(CssFormsGradientSource), nameof(StylesheetProperty));
-
-            var dictionaryAdd = styleProperties.GetType().GetMethod("Add");
-            if (dictionaryAdd == null)
-                return;
-
-            var styleListType = typeof(List<>).MakeGenericType(styleAttributeType);
-            var styleList = (IList)Activator.CreateInstance(styleListType);
-
-            styleList.Add(styleAttributeInstance);
-            dictionaryAdd.Invoke(styleProperties, new object[] { "gradient", styleList });
-        }
-
         public static readonly BindableProperty StylesheetProperty = BindableProperty.Create(
             nameof(Stylesheet), typeof(string), typeof(CssFormsGradientSource));
 
@@ -49,6 +27,21 @@ namespace MagicGradients
         public IEnumerable<LinearGradient> GetGradients()
         {
             return new CssFormsLinearGradientParser().ParseCss(Stylesheet);
+        }
+    }
+
+    [TypeConversion(typeof(CssFormsGradientSource))]
+    public class CssFormsGradientTypeConverter : TypeConverter
+    {
+        public override object ConvertFromInvariantString(string value)
+        {
+            if (string.IsNullOrEmpty(value))
+                throw new InvalidOperationException($"Cannot convert \"{value}\" into {typeof(CssFormsGradientSource)}");
+
+            return new CssFormsGradientSource
+            {
+                Stylesheet = value
+            };
         }
     }
 }
