+++ Playground/Playground/Styles/InlineCssGradient.css
@@ -1,3 +1,4 @@
-﻿CssFormsGradientSource {
-    gradient: linear-gradient(45deg, rgba(171, 212, 83, 0.7) 0%, rgba(171, 212, 83, 0.7) 12.5%,rgba(218, 229, 42, 0.7) 12.5%, rgba(218, 229, 42, 0.7) 25%,rgba(125, 195, 123, 0.7) 25%, rgba(125, 195, 123, 0.7) 37.5%,rgba(194, 221, 63, 0.7) 37.5%, rgba(194, 221, 63, 0.7) 50%,rgba(78, 178, 164, 0.7) 50%, rgba(78, 178, 164, 0.7) 62.5%,rgba(148, 204, 103, 0.7) 62.5%, rgba(148, 204, 103, 0.7) 75%,rgba(241, 238, 22, 0.7) 75%, rgba(241, 238, 22, 0.7) 87.5%,rgba(101, 187, 144, 0.7) 87.5%, rgba(101, 187, 144, 0.7) 100%),linear-gradient(0deg, rgb(73, 182, 19),rgb(231, 253, 15))
+﻿LinearGradientView {
+    background-color: aqua;
+    gradient: linear-gradient(45deg, rgba(171, 212, 83, 0.7) 0%, rgba(171, 212, 83, 0.7) 12.5%,rgba(218, 229, 42, 0.7) 12.5%, rgba(218, 229, 42, 0.7) 25%,rgba(125, 195, 123, 0.7) 25%, rgba(125, 195, 123, 0.7) 37.5%,rgba(194, 221, 63, 0.7) 37.5%, rgba(194, 221, 63, 0.7) 50%,rgba(78, 178, 164, 0.7) 50%, rgba(78, 178, 164, 0.7) 62.5%,rgba(148, 204, 103, 0.7) 62.5%, rgba(148, 204, 103, 0.7) 75%,rgba(241, 238, 22, 0.7) 75%, rgba(241, 238, 22, 0.7) 87.5%,rgba(101, 187, 144, 0.7) 87.5%, rgba(101, 187, 144, 0.7) 100%),linear-gradient(0deg, rgb(73, 182, 19),rgb(231, 253, 15));
 }
